{"types":"C","runs":60}

+++ Problem15.java
@@ -6,8 +6,8 @@ import acm.program.ConsoleProgram;
 import acm.program.*;
 public class Problem15 extends ConsoleProgram{
 	public void run() {
-		int a = readInt("Enter a: ");
-		int b = readInt("Enter b: ");
+		double a = readInt("Enter a: ");
+		double b = readInt("Enter b: ");
 		double average = (a + b) / 2;
 		println("The arithmetic average is "+ average + ".");
 	}
